change container of presentation section signup button
The objective is to be able use the entire body of the button and hit
it wherever one wants. Before it was only possible to click it in the
content part. So the properties were placed directly inside the link tag
to leverage to use of padding and increase the button's hitting space.